# api/openapi.json
{
  "components": {
    "schemas": {
...
      "MeResponse": {
        "description": "The calling user's public profile.",
        "properties": {
          "clerk_user_id": {
            "title": "Clerk User Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "storage": {
            "$ref": "#/components/schemas/StorageResponse"
          }
        },
        "required": [
          "id",
          "clerk_user_id",
          "email",
          "created_at",
          "storage"
        ],
        "title": "MeResponse",
        "type": "object"
      },
...
      "NoticeRequest": {
        "description": "A one-time notice the user has now seen.",
        "properties": {
          "notice": {
            "enum": [
              "first_recording",
              "trial_reminder"
            ],
            "title": "Notice",
            "type": "string"
          }
        },
        "required": [
          "notice"
        ],
        "title": "NoticeRequest",
        "type": "object"
      },
...
      "Problem": {
        "description": "An RFC 9457 problem details body, the shape of every error this API returns.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "description": "`about:blank`, or a problem a client branches on: `urn:crosstune:account-deleted` (401, the account was deleted, so the client drops its local data), `urn:crosstune:premium-required` (403, the addition needs Premium), `urn:crosstune:scan-tune-only` (403, a free account's scans stay on one tune), `urn:crosstune:quota-exceeded` (413), `urn:crosstune:file-too-large` (413).",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "detail"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "ScanUploadSlotRequest": {
        "description": "What the client is about to upload.",
        "properties": {
          "bytes": {
            "exclusiveMinimum": 0.0,
            "title": "Bytes",
            "type": "integer"
          },
          "content_type": {
            "const": "image/jpeg",
            "title": "Content Type",
            "type": "string"
          }
        },
        "required": [
          "bytes",
          "content_type"
        ],
        "title": "ScanUploadSlotRequest",
        "type": "object"
      },
...
      "SignedUrl": {
        "description": "A presigned URL and when it stops working.",
        "properties": {
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "expires_at"
        ],
        "title": "SignedUrl",
        "type": "object"
      },
...
      "StorageResponse": {
        "description": "How much of the storage quota is in use.",
        "properties": {
          "max_file_bytes": {
            "title": "Max File Bytes",
            "type": "integer"
          },
          "quota_bytes": {
            "title": "Quota Bytes",
            "type": "integer"
          },
          "used_bytes": {
            "title": "Used Bytes",
            "type": "integer"
          }
        },
        "required": [
          "used_bytes",
          "quota_bytes",
          "max_file_bytes"
        ],
        "title": "StorageResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Crosstune API",
    "version": "0.19.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/me": {
      "delete": {
        "description": "Delete the calling user's account, every row it owns, and Clerk's copy.\n\nFiles are removed after the response.",
        "operationId": "delete_me_v1_me_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Delete Me",
        "tags": [
          "users"
        ]
      },
      "get": {
        "description": "The calling user's profile and storage figures.",
        "operationId": "me_v1_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Me",
        "tags": [
          "users"
        ]
      }
    },
    "/v1/me/notices": {
      "post": {
        "description": "Record that the user saw a one-time notice, so no device shows it again.\n\nThe first time is the one kept; a repeat changes nothing.",
        "operationId": "record_notice_v1_me_notices_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoticeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Notice",
        "tags": [
          "users"
        ]
      }
    },
...
    "/v1/scans/{scan_id}/download": {
      "get": {
        "description": "A presigned GET for the image of a ready scan.",
        "operationId": "download_v1_scans__scan_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "scan_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignedUrl"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Download",
        "tags": [
          "scans"
        ]
      }
    },
    "/v1/scans/{scan_id}/upload-slot": {
      "post": {
        "description": "A presigned PUT for one scan's image, once the file cap, the plan, and its quota allow it.\n\nA free plan uploads only to its scan tune, and only scans count toward its quota.",
        "operationId": "upload_slot_v1_scans__scan_id__upload_slot_post",
        "parameters": [
          {
            "in": "path",
            "name": "scan_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scan Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanUploadSlotRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignedUrl"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Upload Slot",
        "tags": [
          "scans"
        ]
      }
    },
    "/v1/scans/{scan_id}/uploaded": {
      "post": {
        "description": "Confirm the image landed and mark the scan ready. Repeating the call changes nothing.",
        "operationId": "upload_finished_v1_scans__scan_id__uploaded_post",
        "parameters": [
          {
            "in": "path",
            "name": "scan_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Upload Finished",
        "tags": [
          "scans"
        ]
      }
    },
...
  }
}